Create caption_generator.py
To generate captions, first you’ll create a caption generator. This caption generator utilizes beam search to improve the quality of sentences generated.
At each iteration, the generator passes the previous state of the LSTM (initial state is the image embedding) and previous sequence to generate the next softmax vector.
The top N most probable candidates are kept and utilized in the next inference step. This process continues until either the max sentence length is reached or all sentences have generated the end-of-sentence token.